feature(auth): 🔒 Add in App Engine based user authentication tools.
The old appengine APIs for Python were removed in the new App Engine
generations.

This commit reintroduces the Users API (which is based on request
headers brought in by IAP). This has been put into a `contrib` folder
which will contain other tools that depend on external dependencies.

We are also setting up the base testing infrastructure here.