fix(xsrf)
Remove using the Users class from XSRF as this causes errors. (Users can only be used when IAP is enabled as the App Engine headers are not sent otherwise)

Instead we use a random string generated for the session.

Also updating the app factory to set a Secret Key (needed for Flask Sessions)

Updating the readme to recognise these changes