fix(CSP Headers): CSP Headers are invalid - remove offencing characters.
CSP headers are invalid due to excess quotes and unsafe base64 values.

Whilst base64 values are recommended for CSP NONCE values Python
adds extra "=" signs onto the value as padding. It is necessary to
remove these and provide URL Safe values.

As CSP NONCE values are primarily for allowing scripts through with
the NONCE, this has been added as a setting as well to make it more
accessible.