fix(tasks): Strip off starting slashes from given routes.
Generating a path with os.path.join causes errors when the given route
starts with /. This is because os.path.join assumes it is from the root
and therefore ignores all preceding elements - essentially removing the
url_prefix from the generated path.

This commit strips out starting slashes to avoid this issue.