Use Flask's config logic to load settings, and save SECRET_KEY in the datastore.
Flask has a thing for loading app settings from objects and files. These are
then put into `app.config`. This is nicer than having a custom settings module
thing because other things in the Flask ecosystem can use the app config.

At startup, the SECRET_KEY is stored in the datastore (if it doesn't already
exist) so you don't need to add it to your source code.