Add DatastoreEmulator to help start/stop the datastore emulator.
This works as a context-manager:

    import secure_scaffold.utils

    with secure_scaffold.utils.DatastoreEmulator():
        # Do datastore query stuff
        pass

It accepts arguments to tweak the behaviour of the emulator, e.g. to
use an in-memory database and to change the host/port for binding.

I also made changes to how configuration can be loaded, to make life
easier when testing things because it can be annoying to setup the
database every time. This could be simpler.