Move "secure_scaffold" to "src/securescaffold".
Underscores in Python package names are not ideal. Having the package
in the "src" directory instead of the repo root avoids confusion when
importing the package from a checkout of the repo and you aren't sure
if Python imported the package from the environment site-packages or
from the source tree.

Lots of other things to get this closer to being ready for review,
and working with tests running with Python 3.8.